HSEARCH-3437 Move DslExtensionState to SPI
Because it is used by other modules
