+++ engine/src/main/java/org/hibernate/search/engine/common/dsl/spi/DslExtensionState.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.engine.common.dsl.impl;
+package org.hibernate.search.engine.common.dsl.spi;
 
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
