+++ engine/src/test/java/org/hibernate/search/engine/common/dsl/impl/DslExtensionStateTest.java
@@ -12,6 +12,7 @@ import java.util.Collections;
 import java.util.Optional;
 import java.util.function.Function;
 
+import org.hibernate.search.engine.common.dsl.spi.DslExtensionState;
 import org.hibernate.search.util.SearchException;
 
 import org.junit.Rule;
